backend/dmatat/requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2==2.8.6
uWSGI>=2.0.18,<2.1
textblob==0.15.3